backend/requirements.txt
anyio==4.9.0
asgiref==3.7.2
beautifulsoup4==4.13.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
dadata==24.10.0
defusedxml==0.8.0rc2
Django==5.0.3
django-admin-interface==0.30.0
django-admin-sortable2==2.2.1
django-colorfield==0.14.0
django-cors-headers==4.3.1
django-js-asset==2.2.0
django-mptt==0.16.0
django-rest-framework==0.1.0
django-solo==2.2.0
django-streamfield @ file:///app/packages/django_streamfield-2.3.2-py3-none-any.whl#sha256=c68becde07b91784761634d1e68dd4186777fc6ae5f9d6f26cb7358200940eed
django-templated-mail==1.1.1
django-tinymce==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
gunicorn==22.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.6
lxml==6.0.1
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.2
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.4.0
setuptools==69.5.1
sniffio==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.3
soupsieve==2.6
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.13.2
tzdata==2024.1
urllib3==2.2.1
wheel==0.43.0
